Refactor/is 2927/fix typos and engine naming (#4)
* chore: fix typo in default engine filename

* refactor: rename Local/Remote type names (Current/Latest)

* Update src/backend/electron/engineAndVvppController.ts

* Update src/backend/electron/engineAndVvppController.ts
