+++ tests/unit/domain/defaultEngine/latetDefaultEngine.spec.ts
@@ -1,6 +1,6 @@
 import { vi, expect, test } from "vitest";
 import latestDefaultEngineInfos from "./latestDefaultEngineInfos.json";
-import { fetchLatestDefaultEngineInfo } from "@/domain/defaultEngine/latetDefaultEngine";
+import { fetchLatestDefaultEngineInfo } from "@/domain/defaultEngine/latestDefaultEngine";
 
 test("fetchLatestDefaultEngineInfo", async () => {
   // テスト用のjsonファイルでfetchをモックする
